refactor(memory): 将 _now()、MemoryItem、通用查询方法上提到 BaseMemoryManager
- 新增 _now() 模块级函数到 base.py
- MemoryItem 数据模型从 yaml.py 移至 base.py（存储无关）
- 新增 _load_all() / _save_all() 抽象原语
- show() / show_description_history() / get_memories_grouped() 在基类实现，子类继承
- _generate_id() 在基类提供默认实现（secrets.token_hex(4)）
- yaml.py 精简：移除重复代码，重命名 _read_all/_write_all → _load_all/_save_all
